moved getUsers outside FraHome
