+++ src/client/pages/AssessmentHome/FraHome/FraHome.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { Navigate, NavLink, Route, Routes } from 'react-router-dom'
 
@@ -6,13 +6,9 @@ import classNames from 'classnames'
 
 import { AssessmentHomeRouteNames, ClientRoutes } from '@meta/app'
 import { Areas } from '@meta/area'
-import { Authorizer } from '@meta/user'
 
-import { useAppDispatch } from '@client/store'
-import { useAssessment, useCycle } from '@client/store/assessment'
-import { UserManagementActions } from '@client/store/ui/userManagement'
-import { useUser } from '@client/store/user'
 import { useCountryIso } from '@client/hooks'
+import useGetUsers from '@client/hooks/useGetUsers'
 import User from '@client/pages/User'
 
 import { useSections } from './hooks/useSections'
@@ -22,23 +18,12 @@ import SelectedCountries from './SelectedCountries'
 
 const FraHome: React.FC = () => {
   const { i18n } = useTranslation()
-  const dispatch = useAppDispatch()
   const countryIso = useCountryIso()
-  const assessment = useAssessment()
-  const cycle = useCycle()
-  const user = useUser()
   const sections = useSections()
+  useGetUsers()
 
   // tabs are available when user is logged-in and selected area is country
   const displayTabs = sections.length > 1 && Areas.isISOCountry(countryIso)
-
-  useEffect(() => {
-    if (Authorizer.canViewUsers({ countryIso, assessment, cycle, user })) {
-      dispatch(
-        UserManagementActions.getUsers({ countryIso, assessmentName: assessment.props.name, cycleName: cycle.name })
-      )
-    }
-  }, [countryIso, cycle, assessment, user, dispatch])
 
   return (
     <>
